+++ src/index.d.ts
@@ -1,6 +1,0 @@
-export declare function updraftCliProgram(args: Array<string>): Promise<number>
-
-export declare function updraftProgram(
-	args: Array<string>,
-	usageInstructionsReminder?: string,
-): Promise<number>